rename and reorg tweakpane to debugprovider in preparation for adding/replacing leva
