+++ libraries/commons/debug/src/index.ts
@@ -1,5 +1,5 @@
-export { TweakpaneContext } from "./lib/tweakpane/TweakpaneContext";
-export { TweakpaneProvider } from "./lib/tweakpane/TweakpaneProvider";
-export { useTweakpaneCtx } from "./lib/tweakpane/hooks/useTweakpaneContext";
+export { DebugUiContext } from "./lib/ui/DebugUiContext";
+export { DebugUiProvider } from "./lib/ui/DebugUiProvider";
+export { useTweakpaneCtx } from "./lib/ui/tweakpane/useTweakpaneCtx";
 
 export { useValueChange } from "./lib/hooks/useValueChange";
