--- libraries/commons/debug/src/lib/tweakpane/TweakpaneContext.tsx
@@ -1,13 +1,0 @@
-import React, { createContext } from "react";
-
-import { type Pane } from "tweakpane";
-
-interface TweakpaneContextType {
-  paneRef: React.RefObject<Pane | null>;
-}
-
-export const TweakpaneContext = createContext<TweakpaneContextType | null>(
-  null
-);
-
-export default TweakpaneContext;
